+++ unirest/src/main/java/kong/unirest/json/CoreFactory.java
@@ -1,0 +1,109 @@
+/**
+ * The MIT License
+ *
+ * Copyright for portions of unirest-java are held by Kong Inc (c) 2013.
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining
+ * a copy of this software and associated documentation files (the
+ * "Software"), to deal in the Software without restriction, including
+ * without limitation the rights to use, copy, modify, merge, publish,
+ * distribute, sublicense, and/or sell copies of the Software, and to
+ * permit persons to whom the Software is furnished to do so, subject to
+ * the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be
+ * included in all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+ * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+ * LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+ * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+ * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+
+package kong.unirest.json;
+
+import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
+import com.google.gson.TypeAdapter;
+import com.google.gson.internal.LinkedTreeMap;
+import com.google.gson.stream.JsonReader;
+import com.google.gson.stream.JsonToken;
+import com.google.gson.stream.JsonWriter;
+
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+
+class CoreFactory {
+    private static final TypeAdapter<Object> ADAPTER = new JavaTypeAdapter();
+
+    static GsonBuilder getCore() {
+        return new GsonBuilder()
+                .disableHtmlEscaping()
+                .serializeNulls()
+                .registerTypeAdapter(Map.class, ADAPTER)
+                .registerTypeAdapter(List.class, ADAPTER);
+    }
+
+    static class JavaTypeAdapter extends TypeAdapter<Object> {
+
+        private final TypeAdapter<Object> delegate = new Gson().getAdapter(Object.class);
+
+        @Override
+        public void write(JsonWriter out, Object value) throws IOException {
+            delegate.write(out, value);
+        }
+
+        @Override
+        public Object read(JsonReader in) throws IOException {
+            JsonToken token = in.peek();
+            switch (token) {
+                case BEGIN_ARRAY:
+                    List<Object> list = new ArrayList<>();
+                    in.beginArray();
+                    while (in.hasNext()) {
+                        list.add(read(in));
+                    }
+                    in.endArray();
+                    return list;
+
+                case BEGIN_OBJECT:
+                    Map<String, Object> map = new LinkedTreeMap<>();
+                    in.beginObject();
+                    while (in.hasNext()) {
+                        map.put(in.nextName(), read(in));
+                    }
+                    in.endObject();
+                    return map;
+
+                case STRING:
+                    return in.nextString();
+
+                case NUMBER:
+                    String n = in.nextString();
+                    if (n.indexOf('.') != -1) {
+                        return Double.parseDouble(n);
+                    }
+                    long l = Long.parseLong(n);
+                    if(l < Integer.MAX_VALUE){
+                        return (int)l;
+                    }
+                    return l;
+
+                case BOOLEAN:
+                    return in.nextBoolean();
+
+                case NULL:
+                    in.nextNull();
+                    return null;
+
+                default:
+                    throw new IllegalStateException();
+            }
+        }
+    }
+}